// src/interfaces/common/product-purchase-log.interface.ts
export interface ProductPurchaseLog {
  _id?: string;
  deletedAt?: string;
  deleteYear?: string;
  deleteMonth?: string;
  deleteDateString?: string;
  deletedBy?: string;
  product?: any;
  previousQuantity?: number;
  shop?: any;
  salesman?: any;
  createTime?: any;
  vendor?: any;
  note?: any;
  imei?: any;
  updatedQuantity?: number;
  createdAtString?: string;
  updatedAtString?: string;
  month?: number;
  year?: number;
}